DownloadHandler.js
- Load index.js as a module
- Move download handling into separate file
