+++ index.html
@@ -38,6 +38,6 @@
       <ul id="email-list"></ul>
     </div>
 
-    <script src="index.js"></script>
+    <script src="index.js" type="module"></script>
   </body>
 </html>
